refactored radio style
