+++ Sources/Playbook/Components/Radio/PBRadio.swift
@@ -51,113 +51,31 @@ public struct PBRadio: View {
     switch orientation {
     case .vertical:
       VStack(alignment: .leading, spacing: spacing) {
-        ForEach(items) { item in
-          Button(item.title) {
-            selectedItem = item
-          }
-          .buttonStyle(
-            PBRadioButtonStyle(
-              subtitle: item.subtitle,
-              isSelected: selectedItem?.id == item.id,
-              textAlignment: textAlignment,
-              padding: padding,
-              errorState: errorState
-            )
-          )
-        }
+        itemView
       }
 
     case .horizontal:
       HStack(spacing: spacing) {
-        ForEach(items) { item in
-          Button(item.title) {
-            selectedItem = item
-          }
-          .buttonStyle(
-            PBRadioButtonStyle(
-              subtitle: item.subtitle,
-              isSelected: selectedItem?.id == item.id,
-              textAlignment: textAlignment,
-              padding: padding,
-              errorState: errorState
-            )
-          )
-        }
+        itemView
       }
     }
   }
-}
 
-// MARK: - PBRadioButtonStyle
-
-public struct PBRadioButtonStyle: ButtonStyle {
-  private let subtitle: String
-  private let isSelected: Bool
-  private let textAlignment: Orientation
-  private let padding: CGFloat
-  private let errorState: Bool
-
-  private var borderColor: Color {
-    if errorState {
-      return isSelected ? .pbPrimary : .red
-    } else {
-      return isSelected ? .pbPrimary : .border
-    }
-  }
-
-  private var lineWidth: CGFloat {
-    isSelected ? 6 : 2
-  }
-
-  public init(
-    subtitle: String?,
-    isSelected: Bool,
-    textAlignment: Orientation,
-    padding: CGFloat,
-    errorState: Bool
-  ) {
-    self.isSelected = isSelected
-    self.subtitle = subtitle ?? ""
-    self.textAlignment = textAlignment
-    self.padding = padding
-    self.errorState = errorState
-  }
-
-  public func makeBody(configuration: Configuration) -> some View {
-    switch textAlignment {
-    case .horizontal:
-      HStack(alignment: .top, spacing: padding) {
-        Circle()
-          .strokeBorder(borderColor, lineWidth: lineWidth)
-          .frame(width: 22, height: 22)
-        VStack(alignment: .leading, spacing: Spacing.xxSmall) {
-          configuration.label
-            .foregroundColor(errorState == true ? .status(.error) : .text(.default))
-            .pbFont(.body())
-            .frame(minHeight: 22)
-          if !subtitle.isEmpty {
-            Text(subtitle)
-              .foregroundColor(.text(.light))
-              .pbFont(.subcaption)
-          }
-        }
+  var itemView: some View {
+    ForEach(items) { item in
+      Button(item.title) {
+        selectedItem = item
       }
-      .contentShape(Rectangle())
-    case .vertical:
-      VStack {
-        Circle()
-          .strokeBorder(borderColor, lineWidth: lineWidth)
-          .frame(width: 22, height: 22)
-        VStack(alignment: .leading, spacing: Spacing.xxSmall) {
-          configuration.label
-            .foregroundColor(errorState == true ? .status(.error) :  .text(.default))
-            .pbFont(.body())
-            .frame(minHeight: 22)
-        }
-      }
-      .contentShape(Rectangle())
+      .buttonStyle(
+        PBRadioButtonStyle(
+          subtitle: item.subtitle,
+          isSelected: selectedItem?.id == item.id,
+          textAlignment: textAlignment,
+          padding: padding,
+          errorState: errorState
+        )
+      )
     }
-
   }
 }
